+++ fractals/src/main/java/me/catmousedog/fractals/paneloperators/fractals/Mandelbrot.java
@@ -3,6 +3,7 @@ package me.catmousedog.fractals.paneloperators.fractals;
 import org.jetbrains.annotations.NotNull;
 
 import me.catmousedog.fractals.data.FractalValue;
+import me.catmousedog.fractals.paneloperators.functions.EscapeAngleFunction;
 import me.catmousedog.fractals.paneloperators.functions.Function;
 import me.catmousedog.fractals.paneloperators.functions.IterativeFunction;
 import me.catmousedog.fractals.paneloperators.functions.NormalizedFunction;
@@ -14,7 +15,7 @@ public class Mandelbrot extends Fractal {
 		super();
 		items = null;
 		functions = new Function[] { new IterativeFunction(this), new NormalizedFunction(this),
-				new PotentialFunction(this) };
+				new PotentialFunction(this), new EscapeAngleFunction(this) };
 		function = functions[0];
 		mouse = null;
 	}